requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.4
black==19.10b0
click==7.1.2
dj-database-url==0.5.0
Django==3.0.6
django-pagedown==2.1.2
flake8==3.8.3
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
pathspec==0.8.0
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
python-dotenv==0.13.0
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.12.1
